archive/requirements.txt
yt-dlp
faster-whisper
torch
python-dotenv
mlx-whisper
